service-discovery&registration

+++ NeoBank-Cloud-Config/account/src/main/resources/schema.sql
@@ -10,9 +10,9 @@ CREATE TABLE IF NOT EXISTS `customer` (
     );
 
 CREATE TABLE IF NOT EXISTS `accounts` (
-                                          `customer_id` int NOT NULL,
-                                          `account_number` int AUTO_INCREMENT  PRIMARY KEY,
-                                          `account_type` varchar(100) NOT NULL,
+    `customer_id` int NOT NULL,
+     `account_number` BIGINT AUTO_INCREMENT  PRIMARY KEY,
+     `account_type` varchar(100) NOT NULL,
     `branch_address` varchar(200) NOT NULL,
     `created_at` date NOT NULL,
     `created_by` varchar(20) NOT NULL,
